plitting the data contract and connection
