+++ angular2/index.ts
@@ -1,4 +1,5 @@
-export {DataConnection, DataContract} from './DataObject';
+export {DataContract} from './DataContract';
+export {DataConnection} from './DataConnection';
 export {field} from './field';
 export {register} from '../shared/DataObject';
 export {AuthHandler} from './AuthHandler';
